Implemented linear time graph construction method for GenericGraph.
See issue #160 for background:
https://github.com/JuliaLang/Graphs.jl/issues/160

I introduced a new field in the GenericGraph data type called `indexof` which
is a dictionary mapping V to Int. It gives a quick way to look up the index of
a vertex when encountering a new edge. The hash table look up occurs in
`vertex_index`.

I added corresponding test cases in ../test/graph.jl.